test: fold builtin-skills tests into paths.test.ts
